Added sign in link to home page (#3)
* Added direct links to user authentication

Added links to the homepage linking to signup, login, and sign out. These are only for development.

* Added sign in link to home page

* Added sign in link to home page

Added CSS and html to improve the homepage and add important links to sign up, log in, and sign out. All This information was taken from the Tutorial and will be updated later to reflect the project designs. This is only to make navigation during development easier.

* Corrected bundler info in gemfile.lock

* Added files that are system specific

Files that will be different on everyones machines were added to avoid accidental commiting